+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
-## Version 1.0.0 (unreleased)
+## Version 1.0.1
+
+* __[fix]__ `camelize` defaults to false when not specified (@haines)
+
+## Version 1.0.0
 
 * __[break]__ Removed support for defining types via `type` method (@rstankov)
 * __[break]__ Require `GraphQL::Schema::Resolver` inheritance (@rstankov)